Renamed all environments and experiments files. Moved Benchmark.from_file to Environments.from_file.
