+++ coba/environments/__init__.py
@@ -9,14 +9,15 @@ to do so.
 TODO Add RegressionSimulation
 """
 
-from coba.environments.core import (
-    Context, Action, SimulatedInteraction, Simulation, MemorySimulation, 
-    LambdaSimulation, ClassificationSimulation, CsvSimulation, ArffSimulation, 
-    LibsvmSimulation, ManikSimulation, ValidationSimulation, RegressionSimulation, LoggedInteraction
-)
+from coba.environments.core import Context, Action, SimulatedInteraction, Simulation, LoggedInteraction
 from coba.environments.openml  import OpenmlSource, OpenmlSimulation
 from coba.environments.filters import SimulationFilter, Identity, Shuffle, Take, Sort, Scale, Cycle, Impute
-from coba.environments.pipes   import SimSourceFilters
+from coba.environments.pipes   import EnvironmentPipe
+
+from coba.environments.simulations import (
+    MemorySimulation, LambdaSimulation, ClassificationSimulation, CsvSimulation, 
+    ArffSimulation, LibsvmSimulation, ManikSimulation, ValidationSimulation, RegressionSimulation
+)
 
 __all__ = [
     'Context',
@@ -42,6 +43,6 @@ __all__ = [
     'Scale',
     'Cycle',
     'Impute',
-    'SimSourceFilters',
+    'EnvironmentPipe',
     'LoggedInteraction'
 ]
