+++ coba/tests/test_experiments_core.py
@@ -111,7 +111,7 @@ class Benchmark_Single_Tests(unittest.TestCase):
     def setUpClass(cls) -> None:
         CobaConfig.logger = NullLogger()
         CobaConfig.experiment.processes = 1
-        CobaConfig.experiment.maxtasksperchild = None
+        CobaConfig.experiment.maxtasksperchild = -1
 
     def test_sources(self):
         sim1       = LambdaSimulation(2, lambda i: i, lambda i,c: [0,1,2], lambda i,c,a: cast(float,a))
